Add save/restore logic (#5)
#what Add save and restore logic
#why Important bit of functionality since I can't read the entire track.
Have to play it at least once to gather the data.
This allows someone to do that "once" and then it will be saved and just
be available for future playbacks.

Misc changes:
- Added some rounding (quantization of sorts) to the event times. I
noticed that playback produces slightly different values sometimes. For
this plugin, a few milliseconds one way or the other doesn't matter.
This ensures that the notes don't get duplicated.
- Change the note storage props in the value tree to have a valid xml
name (":midinoteval" instead of just the integer). This allows the
save/restore xml conversion to work without assertion failures. And ...
it's mo' better and correct.

Co-authored-by: Mark Wilkins <markwilkins@github.com>