Improve performance of playback (#6)
Improve the playback performance. Created a sorted vector of time/chord
to be the reference for playback view.
Only refresh it when data is changed and only once per second at most.
This makes a vast improvement in the
playback smoothness. It is able to search efficiently for the position
in the view window and quickly grab those
notes. 

Changes included and related:
- Put the update of the static view in the timer callback in the editor
- Important note: I removed the "show all chords" debug information that
was using the getallnotesonattime ... that was very costly and caused
delays and clicks in playback. Probably can remove that method now.
Unless it is useful for testing.
- Track if the midistate is updated or not. That allows me to only
update the static view if needed.

Co-authored-by: Mark Wilkins <markwilkins@github.com>