Make chord view list update more efficient (#9)
Optimize the playback a little bit by keeping the current set of chords
being displayed in the chord clipper class and modifying that on the fly
for the new position. Most of the time, no changes are needed since we
move forward by a few milliseconds or so.

In reality, I'm not sure it is really better. I think it helps in the
case of a large (e.g., 30 second) view window. But capturing those from
the static view is not very expensive anyway.

---------

Co-authored-by: Mark Wilkins <markwilkins@github.com>