Mlw/misc cleanup (#13)
#what Do some basic cleanup of the look and feel. Moved the setting of
that to the plugin component itself. I think that is the correct place.
#what Change the resolution of the "short chord" slider to be .01
(instead of .1). The finer resolution turns out to be very useful for
busy scores.
#what Fix the slider text to be black on white. It was starting out as
white on grey and then would change to black on white if I opened a
second instance of the plugin editor. Weird. Force it to black on white.
#what Clean up readme a bit more. Add an example image to the readme.

---------

Co-authored-by: Mark Wilkins <markwilkins@github.com>