Add a basic installer and build script (#15)
#what Create a basic installer for the plugins
#why If someone wants to use this without building it themselves, this
*might* be easier.

The issue is that I have not signed the installer. I don't have a
developer subscription with Apple and am not inclined to pay the $99 for
this given that I am likely the only user in the world. If, for some
reason, people find it useful and discover it (how?) then I may pony up
for the developer stuff and build a proper installer.

Co-authored-by: Mark Wilkins <markwilkins@github.com>