Support mouse scrollable view window (#19)
#what Add a listener for mouse scroll (wheel) events and move the view
window accordingly.
#why Sometimes I want to see the chords that are not currently visible. 

This ignores the events if currently in playback (no point in fighting
the constant updates with a mouse movement). If it is currently stopped,
then moves the window by the delta of the event, which is treated as a
ratio of the window size.

The behavior depends on the system settings (mouse intertia,
natural/reverse movement). The info provided by JUCE seems to obey those
settings, so this change is super simple: Listen for the event, adjust
the current position by that given proportion of the window. Literally
about 10 lines of code. The test was much harder to write.

Co-authored-by: Mark Wilkins <markwilkins@github.com>